controller fix
